Added popover for quick toggling and access to injections
I was able to pull out the common CSS between the popover and the manager. I also changed the way the manager JS is kept from the global context.

The popover JS directly access the global page context in order to do things more synchronously. I'm sure this is frowned upon, but the code is way easier to understand.